move recovery to own package and refactor
