+++ plugins/out_clp_s3/out_clp_s3.go
@@ -17,6 +17,7 @@ import (
 
 	"github.com/y-scope/fluent-bit-clp/internal/outctx"
 	"github.com/y-scope/fluent-bit-clp/plugins/out_clp_s3/flush"
+	"github.com/y-scope/fluent-bit-clp/plugins/out_clp_s3/recovery"
 )
 
 const s3PluginName = "out_clp_s3"
@@ -52,7 +53,7 @@ func FLBPluginInit(plugin unsafe.Pointer) int {
 
 	log.Printf("[%s] Init called for id: %s", s3PluginName, outCtx.Config.Id)
 
-	err = flush.RecoverOnStart(outCtx)
+	err = recovery.FlushStores(outCtx)
 	if err != nil {
 		log.Fatalf("Failed to recover logs stored on disk: %s", err)
 	}
@@ -120,7 +121,7 @@ func FLBPluginExitCtx(ctx unsafe.Pointer) int {
 
 	log.Printf("[%s] Exit called for id: %s", s3PluginName, outCtx.Config.Id)
 
-	err := flush.GracefulExit(outCtx)
+	err := recovery.GracefulExit(outCtx)
 	if err != nil {
 		log.Printf("Failed to exit gracefully")
 	}
